+++ algopy_testing/src/algopy/__init__.py
@@ -1,7 +1,4 @@
-# ruff: noqa: PLC0414
-from collections.abc import Callable
-from typing import ParamSpec, TypeVar
-
+from algopy_testing.decorators.subroutine import subroutine
 from algopy_testing.enums import OnCompleteAction, TransactionType
 from algopy_testing.models import (
     Account,
@@ -16,14 +13,6 @@ from algopy_testing.models import (
 )
 from algopy_testing.primitives import BigUInt, Bytes, String, UInt64
 from algopy_testing.state import GlobalState, LocalState
-
-_P = ParamSpec("_P")
-_R = TypeVar("_R")
-
-
-def subroutine(sub: Callable[_P, _R]) -> Callable[_P, _R]:
-    return sub
-
 
 __all__ = [
     "ARC4Contract",
