+++ algopy_testing/src/algopy_testing/models/itxn.py
@@ -17,13 +17,16 @@ class _ITxn:
                 "No inner transaction found in the context! Use `with algopy_testing_context()` "
                 "to access the context manager."
             )
-        itxn = context.inner_transaction_groups[-1]
+        last_itxn_group = context.inner_transaction_groups[-1]
 
-        value = getattr(itxn, name)
+        if not last_itxn_group:
+            raise ValueError("No inner transaction found in the testing context!")
+
+        last_itxn = last_itxn_group[-1]
+
+        value = getattr(last_itxn, name)
         if value is None:
-            # TODO: to match AVM behaviour ideally all inner transaction fields should
-            #       have defaults where possible
-            raise ValueError(f"'{name}' is not defined for {type(itxn).__name__} ")
+            raise ValueError(f"'{name}' is not defined for {type(last_itxn).__name__} ")
         # mimic the static functions on ITxn with a lambda
         return lambda: value
